Move `FeedImageDataLoaderSpy` to a shared scope to remove duplication
