+++ FeedFetchApp/FeedFetchAppTests/FeedImageDataLoaderCacheDecoratorTests.swift
@@ -73,29 +73,3 @@ private extension FeedImageDataLoaderCacheDecoratorTests {
     
     var anyData: Data { return "anydata".data(using: .utf8)! }
 }
-
-private final class FeedImageDataLoaderSpy: FeedImageDataLoader {
-    private(set) var completions = [(FeedImageDataLoader.Result) -> Void]()
-    private(set) var cancelledUrls = [URL]()
-    
-    func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataTask {
-        completions.append(completion)
-        return Task(callback: { [weak self] in
-            self?.cancelledUrls.append(url)
-        })
-    }
-            
-    func complete(with error: NSError, at index: Int = 0) {
-        completions[index](.failure(error))
-    }
-    
-    func complete(with data: Data, at index: Int = 0) {
-        completions[index](.success(data))
-    }
-    
-    private struct Task: FeedImageDataTask {
-        let callback: () -> Void
-                        
-        func cancel() { callback() }
-    }
-}
