+++ FeedFetchApp/FeedFetchAppTests/FeedImageLoaderWithFallbackCompositeTests.swift
@@ -119,32 +119,3 @@ private extension FeedImageLoaderWithFallbackCompositeTests {
     }
             
 }
-
-// MARK:- FeedImageDataLoaderSpy
-
-private final class FeedImageDataLoaderSpy: FeedImageDataLoader {
-    private(set) var completions = [(FeedImageDataLoader.Result) -> Void]()
-    private(set) var cancelledUrls = [URL]()
-    
-    func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataTask {
-        completions.append(completion)
-        return Task(callback: { [weak self] in
-            self?.cancelledUrls.append(url)
-        })
-    }
-            
-    func complete(with error: NSError, at index: Int = 0) {
-        completions[index](.failure(error))
-    }
-    
-    func complete(with data: Data, at index: Int = 0) {
-        completions[index](.success(data))
-    }
-    
-    private struct Task: FeedImageDataTask {
-        let callback: () -> Void
-                        
-        func cancel() { callback() }
-    }
-}
-
